+++ src/smallvec_v1.rs
@@ -201,31 +201,19 @@ where
     }
 }
 
-macro_rules! impl_try_from_into_buf_trait {
-    ($($size:expr),*) => ($(
-        impl<T> TryFrom<[T; $size]> for SmallVec1<[T; $size]> {
-            type Error = Size0Error;
-            fn try_from(vec: [T; $size]) -> Result<Self, Size0Error> {
-                Self::try_from_buf(vec)
-            }
-        }
-
-        impl<T> TryFrom<SmallVec1<[T; $size]>> for [T; $size] {
-            type Error =  SmallVec1<[T; $size]>;
-            fn try_from(vec: SmallVec1<[T; $size]>) -> Result<Self,  SmallVec1<[T; $size]>> {
-                vec.into_inner()
-            }
-        }
-    )*);
+impl<T, const N: usize> TryFrom<[T; N]> for SmallVec1<[T; N]> {
+    type Error = Size0Error;
+    fn try_from(vec: [T; N]) -> Result<Self, Size0Error> {
+        Self::try_from_buf(vec)
+    }
 }
 
-//FIXME support const_generics feature
-impl_try_from_into_buf_trait!(
-    // values from smallvec crate
-    0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 22, 23, 24, 25,
-    26, 27, 28, 29, 30, 31, 32, 36, 0x40, 0x60, 0x80, 0x100, 0x200, 0x400, 0x600, 0x800, 0x1000,
-    0x2000, 0x4000, 0x6000, 0x8000, 0x10_000, 0x20_000, 0x40_000, 0x60_000, 0x80_000, 0x100_000
-);
+impl<T, const N: usize> TryFrom<SmallVec1<[T; N]>> for [T; N] {
+    type Error = SmallVec1<[T; N]>;
+    fn try_from(vec: SmallVec1<[T; N]>) -> Result<Self, SmallVec1<[T; N]>> {
+        vec.into_inner()
+    }
+}
 
 impl<A> IntoIterator for SmallVec1<A>
 where