+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Added missing LICENSE-MIT,LICENSE-APACHE files. Licensing did not change.
 - Added `from_vec_push` and `from_vec_insert` constructors.
 - Use edition 2021.
+- Impl `TryFrom` for `[T; N]` for `Vec1<T>`/`SmallVec1<T>` using const generic.
 
 ## Version 1.8.0 (21.04.2021)